geometry: a sign kernel, letters built from three points, and the fragment in space [skip ci]
Ordering is what the letters, the figures, the shadows and the solids all need
and the equality fragment does not have, so every one of them was smuggling it
in as a Python "<=". This is the smallest thing that fixes that, and the three
things that sit on it.

THE KERNEL. A rational a is non-negative exactly when it is a sum of four
rational squares -- Lagrange applied to pq after writing a = pq/q^2, which is to
say the Pythagoras number of QQ is 4, and 4 is sharp. So an order fact is an
existentially quantified equation and a certificate is four numbers. Checking
one is an addition and an "=="; finding one is a search full of comparisons, and
the module says so in its first paragraph. Measured: a decision 3.2us, a
certificate 91us.

The count of squares is a dimension: the least k is the least d for which
sqrt(a) is a distance between two points of QQ^d. 2 -> the plane; 3 -> space and
not the plane (sqrt 3, the cube's diagonal); 4 -> neither. It stops at QQ and
refuses outside it, because in QQ(sqrt 3) a sum of squares must be positive
under both embeddings and 4*sqrt(3) is not.

WHERE IT CHANGED AN ANSWER, not only the justification -- four ordering bugs the
audit found, all fixed here:
  * occluded: with the light ON the occluder's line, intersection_ll refuses two
    coincident lines and the code read that as "not occluded". A light along the
    occluder cast no shadow at all. Now decided as an interval overlap.
  * shadow_polygon: exact rational corners were sorted by atan2(float(...)) and
    the shoelace area is order-dependent. Now sorted by a half-plane test and
    the sign of a cross product, with no float.
  * prune: the branch-length test went through a float square root, and that
    boolean deletes a branch of a skeleton. Now exact squares.
  * recover: round(float(Fraction)) for three thresholds could cross a tie.
The alphabet evaluation was re-run after the fixes: 325/338, 202/208, 123/130
against 14/130 -- identical on every condition.

THE LETTERS ARE BUILT, NOT TYPED. From o=(0,0), ex=(1,0), ey=(0,1), eighty-two
steps of mirror and midpoint produce all 85 lattice points, and the program
replays from those three alone; both styles agree with the old tables exactly.
From two points nothing ever leaves their line -- 200 constructions, 5 points.

SPACE. Eight primitives over QQ^3, all rational. From three points nothing ever
leaves their plane: 260 points built, 300 constructions, none off it. Two
corrections to how this repository has been stating the plane theorem: it is
about the equivariance of this list, not arithmetic (rational rotations are
plentiful -- 3/5, 4/5), and the reflection must be in a PLANE, since reflect in
a line is a half turn in three dimensions.

And one thing is free here that the plane paid for: a rational equilateral
triangle. (0,0,0), (1,1,0), (1,0,1), all squared edges 2 -- impossible in the
plane, which is why geometry_quadratic adjoined sqrt 3, and predicted exactly by
the kernel's square count.

Tetrahedron, cube and octahedron: rational, regular, convexity certified face by
face. Icosahedron and dodecahedron: QQ(sqrt 5), at every edge length and
orientation, because a rotation of order five has trace (1 +- sqrt 5)/2 and
Niven's theorem leaves SO(3,QQ) the orders 1,2,3,4,6 and never 5.

  python scripts/run_sign_and_space.py --output reports/sign-and-space
  python -m pytest tests/test_geometry_sign_and_space.py  -> 51 passed
  the four geometry suites together -> 107 passed

+++ math_os_prototype/geometry_reading.py
@@ -204,15 +204,15 @@ def recover(bitmap, *, grid=1, tolerance_divisor=12):
     """
     gauge = measure(bitmap)
     body, skeleton, radius = gauge["body"], gauge["skeleton"], gauge["radius"]
-    junction = max(1, round(float(radius)))
+    junction = max(1, round(radius))            # the exact value, not a double
     nodes, edges = raster.skeleton_graph(skeleton, junction_radius=junction)
-    nodes, edges = raster.prune(nodes, edges, max(2, round(float(3*radius/2))))
+    nodes, edges = raster.prune(nodes, edges, max(2, round(3*radius/2)))
     nodes, edges = raster.dissolve(nodes, edges)
     if not edges:
         raise ValueError("the ink has no stroke")
     x0, y0, x1, y1 = gauge["box"]
     tolerance = max(Fraction(1), (y1-y0)/tolerance_divisor)
-    margin = max(1, round(float(radius)))
+    margin = max(1, round(radius))
 
     # every straight part of every chain, as a line and the node it starts from
     parts = []
